langchain/requirements.txt
Flask
flask-cors
python-dotenv
pandas
langchain
langchain-groq
langchain-core
langchain-community
langchain-chroma
chromadb